+++ app/src/components/element-editor-code-tab/validation-code-tab.tsx
@@ -12,7 +12,7 @@ import {Actions} from '../actions/actions';
 import AddOutlinedIcon from '@mui/icons-material/AddOutlined';
 import {useAppDispatch} from '../../hooks/use-app-dispatch';
 import {SelectElementDialog} from '../../dialogs/select-element-dialog/select-element-dialog';
-import {showSuccessSnackbar} from '../../slices/snackbar-slice';
+import {showErrorSnackbar, showSuccessSnackbar} from '../../slices/snackbar-slice';
 import LocationSearchingIcon from '@mui/icons-material/LocationSearching';
 import {createLowCodeContextType} from '../../utils/create-low-code-context-type';
 import {ReferenceCheck} from './components/reference-check/reference-check';
@@ -20,6 +20,7 @@ import {editor} from 'monaco-editor';
 import {ElementValidationFunction} from '../../models/elements/element-validation-function';
 import {NoCodeEditorWrapper} from './components/no-code-editor-wrapper/no-code-editor-wrapper';
 import Delete from '@aivot/mui-material-symbols-400-outlined/dist/delete/Delete';
+import {copyToClipboardText} from '../../utils/copy-to-clipboard';
 
 const exampleValidationCode = `(function(){
     // Hier kann der Code eingefügt werden, der bestimmt, ob das Element valide ist.
@@ -333,8 +334,13 @@ export function ValidationCodeTab(props: ValidationCodeTabProps) {
 
                         dispatch(showSuccessSnackbar('Element-ID eingefügt'));
                     } else {
-                        navigator.clipboard.writeText(element.id);
-                        dispatch(showSuccessSnackbar('Element-ID kopiert'));
+                        copyToClipboardText(element.id).then((success) => {
+                            if (success) {
+                                dispatch(showSuccessSnackbar('Element-ID kopiert'));
+                            } else {
+                                dispatch(showErrorSnackbar('Element-ID konnte nicht kopiert werden'));
+                            }
+                        });
                     }
 
                     toggleShowElementSelectDialog();