+++ app/src/modules/storage/pages/details/storage-provider-details-page.tsx
@@ -126,7 +126,7 @@ export function StorageProviderDetailsPage(): ReactNode {
                     }}
                     parentLink={{
                         label: 'Liste der Speicheranbieter',
-                        to: '/payment-providers',
+                        to: '/storage-providers',
                     }}
                     entityType={ServerEntityType.StorageProviders}
                 />